+++ src/screens/general/BuyShares.js
@@ -5,7 +5,7 @@ import { PayPalButton } from 'react-paypal-button-v2';
 import SharesProgressBar from '../shared/components/SharesProgressBar';
 import LeftArrow from '../../assets/left_arrow.png';
 import '../../styles/BuyShares.css';
-import { recordSharePayment } from '../../lib/paypalUtils';
+import { recordSharePayment } from '../../lib/paypal/paypal';
 import { refreshUserData } from '../../lib/redux/userData';
 import Constants from '../../constants';
 import LoadingComponent from '../../components/LoadingComponent';
