+++ src/screens/onboarding/steps/PaymentStep.js
@@ -3,7 +3,7 @@ import { PayPalButton } from 'react-paypal-button-v2';
 import {
   calculatePaypalTransactionFee,
   recordSharePayment
-} from '../../../lib/paypalUtils';
+} from '../../../lib/paypal/paypal';
 import Constants from '../../../constants';
 import Tooltip from '../components/Tooltip';
